+++ ReleaseNotes.md
@@ -22,11 +22,12 @@ Additional notes:
 The actual release notes
 
 
-# (Upcoming Release)
+# metasfresh 4.6.6 (Upcoming Release)
 
 ## Features
 
 ## Fixes
+ - 09861 Error ESR Import with Partner that has Value bigger than 1000
  - when creating an accounting fact_acct for a C_BankStatementLine_Ref, we now set the fact_acct's C_BPartner_ID correctly
  - 09643 Zahlung und Zahlung-Zuordnung Datum unterschiedlich (108395556223)
     * when allocating an invoice with a payment, we now use the later date of the two documents in that allocation